requirements.txt
gitpython==2.1.11
click==7.0
inquirer==2.5.1
click-completion==0.5.0
pbr==5.1.1
quart==0.7.2
ipfsapi==0.4.3
toml==0.10.0